Add files via upload
As the real estate becomes a scarce commodity in big cities of the world; renovating old
properties for sale is becoming a thriving business. However, not every property becomes an
attractive portfolio after tons of time, money and effort goes into giving it a second life.
Reasons can be its location , condition being too dilapidated or simply the neighbourhood having
lost it lustre of the olden days. These things are not very difficult to figure out after a thorough
assessment. But by the time that assessment is over; a lot of time , manpower and money is lost,
which makes the overall endeavour at times resulting in crippling loss for this upcoming branch
of alternative business opportunity for real estate industry.
Flagging a property as possibly junk before hand can help businesses prioritise their efforts and
focus them on the more probable successes rather than bogged down with the weight of
unsaleable portfolio .
In this project we will be making use of data from such past operations where many properties
were found to be junk after their purchase for renovations. Our task is to predict if a property is
going to be unfit using its listing details and other features which are part of first preliminary
assessment .